+++ src/TinyGsmClientA7672x.h
@@ -712,7 +712,7 @@ class TinyGsmA7672X : public TinyGsmModem<TinyGsmA7672X>,
     return true;
   }
 
-  bool modemBeginSendImpl(uint16_t len, uint8_t mux) {
+  bool modemBeginSendImpl(size_t len, uint8_t mux) {
     if (!sockets[mux]) return 0;
     bool ssl = sockets[mux]->is_secure;
     if (ssl) {
@@ -725,7 +725,7 @@ class TinyGsmA7672X : public TinyGsmModem<TinyGsmA7672X>,
   // Between the modemBeginSend and modemEndSend, modemSend calls:
   // stream.write(reinterpret_cast<const uint8_t*>(buff), len);
   // stream.flush();
-  int16_t modemEndSendImpl(uint16_t len, uint8_t mux) {
+  size_t modemEndSendImpl(size_t len, uint8_t mux) {
     if (!sockets[mux]) return 0;
     bool ssl = sockets[mux]->is_secure;
