+++ src/TinyGsmClientSIM800.h
@@ -596,7 +596,7 @@ class TinyGsmSim800 : public TinyGsmModem<TinyGsmSim800>,
   // Between the modemBeginSend and modemEndSend, modemSend calls:
   // stream.write(reinterpret_cast<const uint8_t*>(buff), len);
   // stream.flush();
-  int16_t modemEndSendImpl(uint16_t, uint8_t mux) {
+  size_t modemEndSendImpl(size_t, uint8_t mux) {
     if (waitResponse(GF(AT_NL "DATA ACCEPT:")) != 1) { return 0; }
     uint8_t ret_mux = streamGetIntBefore(',');   // check mux
     int16_t sent    = streamGetIntBefore('\n');  // check send length
